Refactors BestFitTeamDisplay
