+++ app/imports/ui/layouts/App.jsx
@@ -37,7 +37,7 @@ import TeamInvitationsPage from '../pages/participant/TeamInvitationsPage';
 import AdminEditTeamPage from '../pages/administrator/AdminEditTeamPage';
 import SideBar from '../components/SideBar';
 import ViewTeamsPage from '../pages/administrator/ViewTeamsPage';
-import BestFitTeamDisplay from '../pages/participant/BestFitTeamDisplay';
+import BestFitTeamDisplay from '../pages/participant/BestFitTeam';
 import UpdateMinorParticipantsCompliant from '../pages/administrator/UpdateMinorParticipantsCompliant';
 import HelpPage from '../pages/HelpPage';
 import InterestedParticipantPage from '../pages/participant/InterestedParticipantPage';
