+++ app/package.json
@@ -68,7 +68,7 @@
     "test-acceptance-development-edge": "testcafe edge tests/*.testcafe.js",
     "test-acceptance-development": "testcafe chrome tests/*.testcafe.js",
     "test-acceptance-development-single": "testcafe chrome tests/*.testcafe.js -t \"Test that sign up and sign out work\"",
-    "test-acceptance-development-single-edge": "testcafe edge tests/*.testcafe.js -t \"Test that SuggestionslistAdmin page renders\"",
+    "test-acceptance-development-single-edge": "testcafe edge tests/*.testcafe.js -t \"Test that ListParticipantsAdmin page renders\"",
     "test-acceptance-ci": "testcafe chrome:headless tests/*.testcafe.js -q --app \"meteor npm run start\"",
     "unimported": "npx unimported",
     "test-all": "meteor reset && meteor npm run lint && meteor npm run unimported && meteor npm run test-unit && meteor npm run test-integration && meteor npm run test-acceptance-ci"
